Deleted uneccesary files and edited files
I deleted the config folder because there was a duplicate and some other
files which were uneccesary.
I edited the routes file and managed to make the url of the website be
viewed like this:
http://localhost/Task3PHP2/Graziella_Spiteri_PHP2/home
instead of this:
http://localhost/Task3PHP2/Graziella_Spiteri_PHP2/profile/home.
After a lot of tries the images in the navigation bar started to be
displayed. All I did was clear browsing settings history and retype the
image link. I've cleared the browsing data various time before and for
some reason the images got displayed after this try.